Implement keyboard navigation for the error panel
Fixes #1717

Enable standard "show_panel" with "focus: True" to open the panel and focus
it.  Our toggle panel takes this arg as well and passes it through.

Generally, you *somehow* focus the panel, then `<up>`/`<down>` will move
to next or previous error transiently, t.i. without storing this move in
Sublime's navigation history.  `<enter>` will commit this new location to
said history, while `<escape>` will discard the previewing and return to the
original location.  (Overall this mimics for example the Goto Symbol quick
panel.)  Note that `<escape>` is not specially bound here, we instead
introspect or observe the standard "hide_command" with "cancel: true" command
and act upon it.

On top of this, we may close the panel only if you opened the panel with
"focus: true".  (T.i. you did not move the focus to an already open panel.)

Further notes:

If the panel has focus:

- Change the "SNAP" value to (basically) 0 to the effect that the line
  gets highlighted only if the cursor is exactly on an error.  (T.i.
  turn off the "snap to nearest value" feature.)

- Change the color of our virtual cursor to get *some* visual feedback.
  (For the panel, we choosed total control over the cursor, so basically
	the user never sees a caret. We use a _selection_ (not a region!) to
  highlight error lines; and completely remove any selection otherwise.)
  Note that we will have a normal caret now if the panel has the focus.
  That's normal behavior.